2.0.1

+++ Development/Development/Code/Common/RainAeternum/UtilityString.cpp
@@ -7,7 +7,7 @@ namespace Rain {
 		static std::wstring ret;
 		
 		buffer = new wchar_t[s.length()];
-		bytes = MultiByteToWideChar(CP_UTF8, 0, s.c_str(), s.length(), buffer, s.length());
+		bytes = MultiByteToWideChar(CP_UTF8, 0, s.c_str(), static_cast<int>(s.length()), buffer, static_cast<int>(s.length()));
 		ret = std::wstring(buffer, bytes);
 		delete[] buffer;
 		return ret;
@@ -18,7 +18,7 @@ namespace Rain {
 		static std::string ret;
 
 		buffer = new char[s.length() * 4];
-		bytes = WideCharToMultiByte(CP_UTF8, 0, s.c_str(), s.length(), buffer, s.length() * 4, NULL, NULL);
+		bytes = WideCharToMultiByte(CP_UTF8, 0, s.c_str(), static_cast<int>(s.length()), buffer, static_cast<int>(s.length() * 4), NULL, NULL);
 		ret = std::string(buffer, bytes);
 		delete[] buffer;
 		return ret;
